+++ prov/efa/src/rxr/rxr_cq.c
@@ -436,34 +436,9 @@ write_err:
 	return 0;
 }
 
-static int rxr_cq_match_recv(struct dlist_entry *item, const void *arg)
-{
-	const struct rxr_pkt_entry *pkt_entry = arg;
-	struct rxr_rx_entry *rx_entry;
-
-	rx_entry = container_of(item, struct rxr_rx_entry, entry);
-
-	return rxr_match_addr(rx_entry->addr, pkt_entry->addr);
-}
-
-static int rxr_cq_match_trecv(struct dlist_entry *item, const void *arg)
-{
-	struct rxr_pkt_entry *pkt_entry = (struct rxr_pkt_entry *)arg;
-	struct rxr_rx_entry *rx_entry;
-	uint64_t match_tag;
-
-	rx_entry = container_of(item, struct rxr_rx_entry, entry);
-
-	match_tag = rxr_get_rts_hdr(pkt_entry->pkt)->tag;
-
-	return rxr_match_addr(rx_entry->addr, pkt_entry->addr) &&
-	       rxr_match_tag(rx_entry->cq_entry.tag, rx_entry->ignore,
-			     match_tag);
-}
-
-static void rxr_cq_post_connack(struct rxr_ep *ep,
-				struct rxr_peer *peer,
-				fi_addr_t addr)
+void rxr_pkt_post_connack(struct rxr_ep *ep,
+			  struct rxr_peer *peer,
+			  fi_addr_t addr)
 {
 	struct rxr_pkt_entry *pkt_entry;
 	ssize_t ret;
@@ -649,230 +624,9 @@ void rxr_cq_handle_rx_completion(struct rxr_ep *ep,
 	return;
 }
 
-ssize_t rxr_cq_recv_shm_large_message(struct rxr_ep *ep, struct rxr_rx_entry *rx_entry)
-{
-	struct rxr_pkt_entry *pkt_entry;
-	struct rxr_rma_context_pkt *rma_context_pkt;
-	struct fi_msg_rma msg;
-	struct iovec msg_iov[RXR_IOV_LIMIT];
-	struct fi_rma_iov rma_iov[RXR_IOV_LIMIT];
-	fi_addr_t src_shm_fiaddr;
-	uint64_t remain_len;
-	struct rxr_peer *peer;
-	int ret, i;
-
-	if (rx_entry->state == RXR_RX_QUEUED_SHM_LARGE_READ)
-		return 0;
-
-	pkt_entry = rxr_pkt_entry_alloc(ep, ep->tx_pkt_shm_pool);
-	assert(pkt_entry);
-
-	pkt_entry->x_entry = (void *)rx_entry;
-	pkt_entry->addr = rx_entry->addr;
-	rma_context_pkt = (struct rxr_rma_context_pkt *)pkt_entry->pkt;
-	rma_context_pkt->type = RXR_RMA_CONTEXT_PKT;
-	rma_context_pkt->version = RXR_PROTOCOL_VERSION;
-	rma_context_pkt->rma_context_type = RXR_SHM_LARGE_READ;
-	rma_context_pkt->tx_id = rx_entry->tx_id;
-
-	peer = rxr_ep_get_peer(ep, rx_entry->addr);
-	src_shm_fiaddr = peer->shm_fiaddr;
-
-	memset(&msg, 0, sizeof(msg));
-
-	remain_len = rx_entry->total_len;
-
-	for (i = 0; i < rx_entry->rma_iov_count; i++) {
-		rma_iov[i].addr = rx_entry->rma_iov[i].addr;
-		rma_iov[i].len = rx_entry->rma_iov[i].len;
-		rma_iov[i].key = 0;
-	}
-
-	/*
-	 * shm provider will compare #bytes CMA copied with total length of recv buffer
-	 * (msg_iov here). If they are not equal, an error is returned when reading shm
-	 * provider's cq. So shrink the total length of recv buffer if applicable
-	 */
-	for (i = 0; i < rx_entry->iov_count; i++) {
-		msg_iov[i].iov_base = (void *)rx_entry->iov[i].iov_base;
-		msg_iov[i].iov_len = (remain_len < rx_entry->iov[i].iov_len) ?
-					remain_len : rx_entry->iov[i].iov_len;
-		remain_len -= msg_iov[i].iov_len;
-		if (remain_len == 0)
-			break;
-	}
-
-	msg.msg_iov = msg_iov;
-	msg.iov_count = rx_entry->iov_count;
-	msg.desc = NULL;
-	msg.addr = src_shm_fiaddr;
-	msg.context = pkt_entry;
-	msg.rma_iov = rma_iov;
-	msg.rma_iov_count = rx_entry->rma_iov_count;
-
-	ret = fi_readmsg(ep->shm_ep, &msg, 0);
-
-	return ret;
-}
-
-void rxr_cq_process_shm_large_message(struct rxr_ep *ep, struct rxr_rx_entry *rx_entry,
-				      struct rxr_rts_hdr *rts_hdr, char *data)
-{
-	struct iovec *iovec_ptr;
-	int ret, i;
-
-	/* get iov_count of sender first */
-	memcpy(&rx_entry->rma_iov_count, data, sizeof(size_t));
-	data += sizeof(size_t);
-
-	iovec_ptr = (struct iovec *)data;
-	for (i = 0; i < rx_entry->rma_iov_count; i++) {
-		iovec_ptr = iovec_ptr + i;
-		rx_entry->rma_iov[i].addr = (intptr_t) iovec_ptr->iov_base;
-		rx_entry->rma_iov[i].len = iovec_ptr->iov_len;
-		rx_entry->rma_iov[i].key = 0;
-	}
-
-	ret = rxr_cq_recv_shm_large_message(ep, rx_entry);
-
-	if (OFI_UNLIKELY(ret)) {
-		if (ret == -FI_EAGAIN) {
-			rx_entry->state = RXR_RX_QUEUED_SHM_LARGE_READ;
-			dlist_insert_tail(&rx_entry->queued_entry,  &ep->rx_entry_queued_list);
-			return;
-		}
-		FI_WARN(&rxr_prov, FI_LOG_CQ,
-			"A large message RMA READ failed over shm provider.\n");
-		if (rxr_cq_handle_rx_error(ep, rx_entry, ret))
-			assert(0 && "failed to write err cq entry");
-	}
-}
-
-char *rxr_cq_read_rts_hdr(struct rxr_ep *ep,
-			  struct rxr_rx_entry *rx_entry,
-			  struct rxr_pkt_entry *pkt_entry)
-{
-	char *data;
-	struct rxr_rts_hdr *rts_hdr = NULL;
-	/*
-	 * Use the correct header and grab CQ data and data, but ignore the
-	 * source_address since that has been fetched and processed already
-	 */
-
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-
-	rx_entry->addr = pkt_entry->addr;
-	rx_entry->tx_id = rts_hdr->tx_id;
-	rx_entry->msg_id = rts_hdr->msg_id;
-	rx_entry->total_len = rts_hdr->data_len;
-	rx_entry->cq_entry.tag = rts_hdr->tag;
-
-	if (rts_hdr->flags & RXR_REMOTE_CQ_DATA) {
-		rx_entry->cq_entry.flags |= FI_REMOTE_CQ_DATA;
-		data = rxr_get_ctrl_cq_pkt(rts_hdr)->data + rts_hdr->addrlen;
-		rx_entry->cq_entry.data =
-				rxr_get_ctrl_cq_pkt(rts_hdr)->hdr.cq_data;
-	} else {
-		rx_entry->cq_entry.data = 0;
-		data = rxr_get_ctrl_pkt(rts_hdr)->data + rts_hdr->addrlen;
-	}
-
-	return data;
-}
-
-int rxr_cq_process_msg_rts(struct rxr_ep *ep,
-			   struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_peer *peer;
-	struct rxr_rts_hdr *rts_hdr;
-	struct dlist_entry *match;
-	struct rxr_rx_entry *rx_entry;
-	char *data;
-	size_t data_size;
-
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-
-	if (rts_hdr->flags & RXR_TAGGED) {
-		match = dlist_find_first_match(&ep->rx_tagged_list,
-					       &rxr_cq_match_trecv,
-					       (void *)pkt_entry);
-	} else {
-		match = dlist_find_first_match(&ep->rx_list,
-					       &rxr_cq_match_recv,
-					       (void *)pkt_entry);
-	}
-
-	if (OFI_UNLIKELY(!match)) {
-		rx_entry = rxr_ep_get_new_unexp_rx_entry(ep, pkt_entry);
-		if (!rx_entry) {
-			FI_WARN(&rxr_prov, FI_LOG_CQ,
-				"RX entries exhausted.\n");
-			efa_eq_write_error(&ep->util_ep, FI_ENOBUFS, -FI_ENOBUFS);
-			return -FI_ENOBUFS;
-		}
-
-		/* we are not releasing pkt_entry here because it will be
-		 * processed later
-		 */
-		pkt_entry = rx_entry->unexp_rts_pkt;
-		rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-		rxr_cq_read_rts_hdr(ep, rx_entry, pkt_entry);
-		return 0;
-	}
-
-	rx_entry = container_of(match, struct rxr_rx_entry, entry);
-	if (rx_entry->rxr_flags & RXR_MULTI_RECV_POSTED) {
-		rx_entry = rxr_ep_split_rx_entry(ep, rx_entry,
-						 NULL, pkt_entry);
-		if (OFI_UNLIKELY(!rx_entry)) {
-			FI_WARN(&rxr_prov, FI_LOG_CQ,
-				"RX entries exhausted.\n");
-			efa_eq_write_error(&ep->util_ep, FI_ENOBUFS, -FI_ENOBUFS);
-			return -FI_ENOBUFS;
-		}
-	}
-
-	rx_entry->state = RXR_RX_MATCHED;
-
-	if (!(rx_entry->fi_flags & FI_MULTI_RECV) ||
-	    !rxr_msg_multi_recv_buffer_available(ep, rx_entry->master_entry))
-		dlist_remove(match);
-
-	peer = rxr_ep_get_peer(ep, pkt_entry->addr);
-	assert(peer);
-
-	data = rxr_cq_read_rts_hdr(ep, rx_entry, pkt_entry);
-	if (peer->is_local && !(rts_hdr->flags & RXR_SHM_HDR_DATA)) {
-		rxr_cq_process_shm_large_message(ep, rx_entry, rts_hdr, data);
-		rxr_pkt_entry_release_rx(ep, pkt_entry);
-		return 0;
-	}
-
-	data_size = rxr_get_rts_data_size(ep, rts_hdr);
-	return rxr_cq_handle_rts_with_data(ep, rx_entry,
-					   pkt_entry, data,
-					   data_size);
-}
-
-static int rxr_cq_process_rts(struct rxr_ep *ep,
-			      struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_rts_hdr *rts_hdr;
-
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-
-	if (rts_hdr->flags & RXR_READ_REQ)
-		return rxr_rma_proc_read_rts(ep, pkt_entry);
-
-	if (rts_hdr->flags & RXR_WRITE)
-		return rxr_rma_proc_write_rts(ep, pkt_entry);
-
-	return rxr_cq_process_msg_rts(ep, pkt_entry);
-}
-
-static int rxr_cq_reorder_msg(struct rxr_ep *ep,
-			      struct rxr_peer *peer,
-			      struct rxr_pkt_entry *pkt_entry)
+int rxr_cq_reorder_msg(struct rxr_ep *ep,
+		       struct rxr_peer *peer,
+		       struct rxr_pkt_entry *pkt_entry)
 {
 	struct rxr_rts_hdr *rts_hdr;
 	struct rxr_pkt_entry *ooo_entry;
@@ -917,8 +671,8 @@ static int rxr_cq_reorder_msg(struct rxr_ep *ep,
 	return 1;
 }
 
-static void rxr_cq_proc_pending_items_in_recvwin(struct rxr_ep *ep,
-						 struct rxr_peer *peer)
+void rxr_cq_proc_pending_items_in_recvwin(struct rxr_ep *ep,
+					  struct rxr_peer *peer)
 {
 	struct rxr_pkt_entry *pending_pkt;
 	struct rxr_rts_hdr *rts_hdr;
@@ -935,8 +689,8 @@ static void rxr_cq_proc_pending_items_in_recvwin(struct rxr_ep *ep,
 		FI_DBG(&rxr_prov, FI_LOG_EP_CTRL,
 		       "Processing msg_id %d from robuf\n", rts_hdr->msg_id);
 
-		/* rxr_cq_process_rts will write error cq entry if needed */
-		ret = rxr_cq_process_rts(ep, pending_pkt);
+		/* rxr_pkt_proc_rts() will write error cq entry if needed */
+		ret = rxr_pkt_proc_rts(ep, pending_pkt);
 		if (OFI_UNLIKELY(ret)) {
 			FI_WARN(&rxr_prov, FI_LOG_CQ,
 				"Error processing msg_id %d from robuf: %s\n",
@@ -1007,69 +761,6 @@ static void rxr_cq_handle_eor(struct rxr_ep *ep,
 	rxr_pkt_entry_release_rx(ep, pkt_entry);
 }
 
-static void rxr_cq_handle_rts(struct rxr_ep *ep,
-			      struct fi_cq_data_entry *comp,
-			      struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_rts_hdr *rts_hdr;
-	struct rxr_peer *peer;
-	int ret;
-
-	peer = rxr_ep_get_peer(ep, pkt_entry->addr);
-	assert(peer);
-
-	if (rxr_env.enable_shm_transfer && peer->is_local) {
-		/* no need to reorder msg for shm_ep
-		 * rxr_cq_process_rts will write error cq entry if needed
-		 */
-		rxr_cq_process_rts(ep, pkt_entry);
-		return;
-	}
-
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-
-	if (ep->core_caps & FI_SOURCE)
-		rxr_cq_post_connack(ep, peer, pkt_entry->addr);
-
-	if (rxr_need_sas_ordering(ep)) {
-		ret = rxr_cq_reorder_msg(ep, peer, pkt_entry);
-		if (ret == 1) {
-			/* Packet was queued */
-			return;
-		} else if (OFI_UNLIKELY(ret == -FI_EALREADY)) {
-			FI_WARN(&rxr_prov, FI_LOG_EP_CTRL,
-				"Invalid msg_id: %" PRIu32
-				" robuf->exp_msg_id: %" PRIu32 "\n",
-			       rts_hdr->msg_id, peer->robuf->exp_msg_id);
-			if (!rts_hdr->addrlen)
-				efa_eq_write_error(&ep->util_ep, FI_EIO, ret);
-			rxr_pkt_entry_release_rx(ep, pkt_entry);
-			return;
-		} else if (OFI_UNLIKELY(ret == -FI_ENOMEM)) {
-			efa_eq_write_error(&ep->util_ep, FI_ENOBUFS, -FI_ENOBUFS);
-			return;
-		} else if (OFI_UNLIKELY(ret < 0)) {
-			FI_WARN(&rxr_prov, FI_LOG_EP_CTRL,
-				"Unknown error %d processing RTS packet msg_id: %"
-				PRIu32 "\n", ret, rts_hdr->msg_id);
-			efa_eq_write_error(&ep->util_ep, FI_EIO, ret);
-			return;
-		}
-
-		/* processing the expected packet */
-		ofi_recvwin_slide(peer->robuf);
-	}
-
-	/* rxr_cq_process_rts will write error cq entry if needed */
-	ret = rxr_cq_process_rts(ep, pkt_entry);
-	if (OFI_UNLIKELY(ret))
-		return;
-
-	/* process pending items in reorder buff */
-	if (rxr_need_sas_ordering(ep))
-		rxr_cq_proc_pending_items_in_recvwin(ep, peer);
-}
-
 static void rxr_cq_handle_connack(struct rxr_ep *ep,
 				  struct fi_cq_data_entry *comp,
 				  struct rxr_pkt_entry *pkt_entry,
@@ -1087,65 +778,6 @@ static void rxr_cq_handle_connack(struct rxr_ep *ep,
 	FI_DBG(&rxr_prov, FI_LOG_CQ,
 	       "CONNACK received from %" PRIu64 "\n", src_addr);
 	rxr_pkt_entry_release_rx(ep, pkt_entry);
-}
-
-int rxr_cq_handle_rts_with_data(struct rxr_ep *ep,
-				struct rxr_rx_entry *rx_entry,
-				struct rxr_pkt_entry *pkt_entry,
-				char *data, size_t data_size)
-{
-	struct rxr_rts_hdr *rts_hdr;
-	int64_t bytes_left, bytes_copied;
-	ssize_t ret;
-
-
-	/* rx_entry->cq_entry.len is total recv buffer size.
-	 * rx_entry->total_len is from rts_hdr and is total send buffer size.
-	 * if send buffer size < recv buffer size, we adjust value of rx_entry->cq_entry.len.
-	 * if send buffer size > recv buffer size, we have a truncated message.
-	 */
-	if (rx_entry->cq_entry.len > rx_entry->total_len)
-		rx_entry->cq_entry.len = rx_entry->total_len;
-
-	bytes_copied = ofi_copy_to_iov(rx_entry->iov, rx_entry->iov_count,
-				       0, data, data_size);
-
-	if (OFI_UNLIKELY(bytes_copied < data_size)) {
-		/* recv buffer is not big enough to hold rts, this must be a truncated message */
-		assert(bytes_copied == rx_entry->cq_entry.len &&
-		       rx_entry->cq_entry.len < rx_entry->total_len);
-		rx_entry->bytes_done = bytes_copied;
-		bytes_left = 0;
-	} else {
-		assert(bytes_copied == data_size);
-		rx_entry->bytes_done = data_size;
-		bytes_left = rx_entry->total_len - data_size;
-	}
-
-	assert(bytes_left >= 0);
-	if (!bytes_left) {
-		/* rxr_cq_handle_rx_completion() releases pkt_entry, thus
-		 * we do not release it here.
-		 */
-		rxr_cq_handle_rx_completion(ep, pkt_entry, rx_entry);
-		rxr_msg_multi_recv_free_posted_entry(ep, rx_entry);
-		rxr_release_rx_entry(ep, rx_entry);
-		return 0;
-	}
-
-#if ENABLE_DEBUG
-	dlist_insert_tail(&rx_entry->rx_pending_entry, &ep->rx_pending_list);
-	ep->rx_pending++;
-#endif
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-	rx_entry->state = RXR_RX_RECV;
-	if (rts_hdr->flags & RXR_CREDIT_REQUEST)
-		rx_entry->credit_request = rts_hdr->credit_request;
-	else
-		rx_entry->credit_request = rxr_env.tx_min_credits;
-	ret = rxr_ep_post_ctrl_or_queue(ep, RXR_RX_ENTRY, rx_entry, RXR_CTS_PKT, 0);
-	rxr_pkt_entry_release_rx(ep, pkt_entry);
-	return ret;
 }
 
 int rxr_cq_handle_pkt_with_data(struct rxr_ep *ep,
@@ -1413,7 +1045,7 @@ void rxr_cq_handle_pkt_recv_completion(struct rxr_ep *ep,
 
 	switch (rxr_get_base_hdr(pkt_entry->pkt)->type) {
 	case RXR_RTS_PKT:
-		rxr_cq_handle_rts(ep, cq_entry, pkt_entry);
+		rxr_pkt_handle_rts_recv(ep, pkt_entry);
 		return;
 	case RXR_EOR_PKT:
 		rxr_cq_handle_eor(ep, cq_entry, pkt_entry);
