+++ prov/efa/src/rxr/rxr_ep.c
@@ -876,185 +876,6 @@ void rxr_ep_init_connack_pkt_entry(struct rxr_ep *ep,
 	pkt_entry->addr = addr;
 }
 
-/* RTS related functions */
-char *rxr_pkt_init_rts_base_hdr(struct rxr_ep *ep,
-				struct rxr_tx_entry *tx_entry,
-				struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_rts_hdr *rts_hdr;
-	struct rxr_peer *peer;
-	char *src;
-
-	rts_hdr = (struct rxr_rts_hdr *)pkt_entry->pkt;
-	peer = rxr_ep_get_peer(ep, tx_entry->addr);
-
-	rts_hdr->type = RXR_RTS_PKT;
-	rts_hdr->version = RXR_PROTOCOL_VERSION;
-	rts_hdr->tag = tx_entry->tag;
-
-	rts_hdr->data_len = tx_entry->total_len;
-	rts_hdr->tx_id = tx_entry->tx_id;
-	rts_hdr->msg_id = tx_entry->msg_id;
-	/*
-	 * Even with protocol versions prior to v3 that did not include a
-	 * request in the RTS, the receiver can test for this flag and decide if
-	 * it should be used as a heuristic for credit calculation. If the
-	 * receiver is on <3 protocol version, the flag and the request just get
-	 * ignored.
-	 */
-	rts_hdr->flags |= RXR_CREDIT_REQUEST;
-	rts_hdr->credit_request = tx_entry->credit_request;
-
-	if (tx_entry->fi_flags & FI_REMOTE_CQ_DATA) {
-		rts_hdr->flags = RXR_REMOTE_CQ_DATA;
-		pkt_entry->pkt_size = RXR_CTRL_HDR_SIZE;
-		rxr_get_ctrl_cq_pkt(rts_hdr)->hdr.cq_data =
-			tx_entry->cq_entry.data;
-		src = rxr_get_ctrl_cq_pkt(rts_hdr)->data;
-	} else {
-		rts_hdr->flags = 0;
-		pkt_entry->pkt_size = RXR_CTRL_HDR_SIZE_NO_CQ;
-		src = rxr_get_ctrl_pkt(rts_hdr)->data;
-	}
-
-	if (tx_entry->cq_entry.flags & FI_TAGGED)
-		rts_hdr->flags |= RXR_TAGGED;
-
-	rts_hdr->addrlen = 0;
-	if (OFI_UNLIKELY(peer->state != RXR_PEER_ACKED)) {
-		/*
-		 * This is the first communication with this peer on this
-		 * endpoint, so send the core's address for this EP in the RTS
-		 * so the remote side can insert it into its address vector.
-		 */
-		rts_hdr->addrlen = ep->core_addrlen;
-		rts_hdr->flags |= RXR_REMOTE_SRC_ADDR;
-		memcpy(src, ep->core_addr, rts_hdr->addrlen);
-		src += rts_hdr->addrlen;
-		pkt_entry->pkt_size += rts_hdr->addrlen;
-	}
-
-	return src;
-}
-
-ssize_t rxr_pkt_init_rts(struct rxr_ep *ep,
-			 struct rxr_tx_entry *tx_entry,
-			 struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_peer *peer;
-	struct rxr_rts_hdr *rts_hdr;
-	char *data, *src;
-	uint64_t data_len;
-	size_t mtu = ep->mtu_size;
-
-	if (tx_entry->op == ofi_op_read_req)
-		return rxr_rma_init_read_rts(ep, tx_entry, pkt_entry);
-
-	src = rxr_pkt_init_rts_base_hdr(ep, tx_entry, pkt_entry);
-	if (tx_entry->op == ofi_op_write)
-		src = rxr_pkt_init_rts_rma_hdr(ep, tx_entry, pkt_entry, src);
-
-	peer = rxr_ep_get_peer(ep, tx_entry->addr);
-	assert(peer);
-	data = src;
-	rts_hdr = rxr_get_rts_hdr(pkt_entry->pkt);
-	if (rxr_env.enable_shm_transfer && peer->is_local) {
-		rts_hdr->flags |= RXR_SHM_HDR;
-		/* will be sent over shm provider */
-		if (tx_entry->total_len <= rxr_env.shm_max_medium_size) {
-			data_len = ofi_copy_from_iov(data, rxr_env.shm_max_medium_size,
-						     tx_entry->iov, tx_entry->iov_count, 0);
-			assert(data_len == tx_entry->total_len);
-			rts_hdr->flags |= RXR_SHM_HDR_DATA;
-			pkt_entry->pkt_size += data_len;
-		} else {
-			/* rendezvous protocol
-			 * place iov_count first, then local iov
-			 */
-			memcpy(data, &tx_entry->iov_count, sizeof(size_t));
-			data += sizeof(size_t);
-			pkt_entry->pkt_size += sizeof(size_t);
-			memcpy(data, tx_entry->iov, sizeof(struct iovec) * tx_entry->iov_count);
-			pkt_entry->pkt_size += sizeof(struct iovec) * tx_entry->iov_count;
-		}
-	} else {
-		/* will be sent over efa provider */
-		data_len = ofi_copy_from_iov(data, mtu - pkt_entry->pkt_size,
-					     tx_entry->iov, tx_entry->iov_count, 0);
-		assert(data_len == rxr_get_rts_data_size(ep, rts_hdr));
-		pkt_entry->pkt_size += data_len;
-	}
-
-	assert(pkt_entry->pkt_size <= mtu);
-	pkt_entry->addr = tx_entry->addr;
-	pkt_entry->x_entry = (void *)tx_entry;
-	return 0;
-}
-
-uint64_t rxr_get_rts_data_size(struct rxr_ep *ep,
-			       struct rxr_rts_hdr *rts_hdr)
-{
-	/*
-	 * read RTS contain no data, because data is on remote EP.
-	 */
-	if (rts_hdr->flags & RXR_READ_REQ)
-		return 0;
-
-	if (rts_hdr->flags & RXR_SHM_HDR)
-		return (rts_hdr->flags & RXR_SHM_HDR_DATA) ? rts_hdr->data_len : 0;
-
-	size_t max_payload_size;
-
-	if (rts_hdr->flags & RXR_REMOTE_CQ_DATA)
-		max_payload_size = ep->mtu_size - RXR_CTRL_HDR_SIZE;
-	else
-		max_payload_size = ep->mtu_size - RXR_CTRL_HDR_SIZE_NO_CQ;
-
-	if (rts_hdr->flags & RXR_REMOTE_SRC_ADDR)
-		max_payload_size -= rts_hdr->addrlen;
-
-	if (rts_hdr->flags & RXR_WRITE)
-		max_payload_size -= rts_hdr->rma_iov_count *
-					sizeof(struct fi_rma_iov);
-
-	return (rts_hdr->data_len > max_payload_size)
-		? max_payload_size : rts_hdr->data_len;
-}
-
-void rxr_pkt_handle_rts_sent(struct rxr_ep *ep, struct rxr_pkt_entry *pkt_entry)
-{
-	struct rxr_peer *peer;
-	struct rxr_tx_entry *tx_entry;
-	size_t data_sent;
-
-	tx_entry = (struct rxr_tx_entry *)pkt_entry->x_entry;
-
-	peer = rxr_ep_get_peer(ep, pkt_entry->addr);
-	assert(peer);
-	if (tx_entry->op == ofi_op_read_req) {
-		tx_entry->bytes_sent = 0;
-		tx_entry->state = RXR_TX_WAIT_READ_FINISH;
-		return;
-	}
-
-	data_sent = rxr_get_rts_data_size(ep, rxr_get_rts_hdr(pkt_entry->pkt));
-
-	tx_entry->bytes_sent += data_sent;
-
-	if ((rxr_env.enable_shm_transfer && peer->is_local) ||
-	    !(efa_mr_cache_enable && tx_entry->total_len > data_sent))
-		return;
-
-	/*
-	 * Register the data buffers inline only if the application did not
-	 * provide a descriptor with the tx op
-	 */
-	if (rxr_ep_mr_local(ep) && !tx_entry->desc[0])
-		rxr_inline_mr_reg(rxr_ep_domain(ep), tx_entry);
-
-	return;
-}
-
 int rxr_ep_init_ctrl_pkt(struct rxr_ep *rxr_ep, int entry_type, void *x_entry,
 			 int ctrl_type, struct rxr_pkt_entry *pkt_entry)
 {
@@ -2106,7 +1927,7 @@ void rxr_ep_progress_internal(struct rxr_ep *ep)
 					       rx_entry->queued_ctrl.type,
 					       rx_entry->queued_ctrl.inject);
 		else if (rx_entry->state == RXR_RX_QUEUED_SHM_LARGE_READ)
-			ret = rxr_cq_recv_shm_large_message(ep, rx_entry);
+			ret = rxr_pkt_post_shm_rndzv_read(ep, rx_entry);
 		else
 			ret = rxr_ep_send_queued_pkts(ep,
 						      &rx_entry->queued_pkts);
